Pre worlds v2

+++ src/org/usfirst/frc/team4028/robot/utilities/ShooterTableEntry.java
@@ -3,7 +3,7 @@ package org.usfirst.frc.team4028.robot.utilities;
 public class ShooterTableEntry { 
 	// define class level working variables
 	public final int Index;
-	public final int DistanceInInches;
+	public final double DistanceInInches;
 	public final String Description;
 	public final double SliderPosition;
 	public final int Stg1MotorRPM;
@@ -13,7 +13,7 @@ public class ShooterTableEntry {
 	//============================================================================================
 	// constructors follow
 	//============================================================================================
-	public ShooterTableEntry(int index, int distanceInInches, String description, double sliderPosition, int stg1MotorRPM, int stg2MotorRPM, boolean isDefault) {
+	public ShooterTableEntry(int index, double distanceInInches, String description, double sliderPosition, int stg1MotorRPM, int stg2MotorRPM, boolean isDefault) {
 		Index = index;
 		DistanceInInches = distanceInInches;
 		Description = description;
